+++ pfio/MultiGroupServer.F90
@@ -87,12 +87,14 @@ module pFIO_MultiGroupServerMod
 
 contains
 
-   function new_MultiGroupServer(server_comm, port_name, nwriter_per_node, rc) result(s)
+   function new_MultiGroupServer(server_comm, port_name, nwriter_per_node, with_profiler, rc) result(s)
       type (MultiGroupServer) :: s
       integer, intent(in) :: server_comm
       character(*), intent(in) :: port_name
       integer, intent (in) :: nwriter_per_node
+      logical, optional, intent(in) :: with_profiler
       integer, optional, intent(out) :: rc
+
       integer :: s_rank, s_size
       integer :: ierror, status, local_rank
       type (SimpleCommSplitter) :: splitter
@@ -128,7 +130,7 @@ contains
       s%I_am_back_root = .false.
       if (index(s_name, 'o_server_front') /=0) then
          s%front_comm = s_comm%get_subcommunicator()
-         call s%init(s%front_comm, s_name)
+         call s%init(s%front_comm, s_name, with_profiler = with_profiler)
          s%port_name = trim(port_name)
          call MPI_Comm_rank(s%front_comm, local_rank, ierror)
          if (s_rank == 0) then